Revert "move to perf test suite"
This reverts commit 4ef2f519ed35cfdc02862600befb5ac6868b271a.
